requirements.txt
openai==1.98.0
python-dotenv==1.1.1
pyaudio==0.2.11
keyboard==0.13.5
pydub==0.25.1
numpy==1.24.3